Improve items_in_stock updating when using PLATA_STOCK_TRACKING
Stock transactions of type PAYMENT_PROCESS_RESERVATION should not
influence the items_in_stock field on the product itself, because
if we get a payment failure and the 15 minutes are over we still
have the lower items_in_stock amount in the database until the
next stock transaction for the product is saved or deleted.

The payment process reservation stock transactions are only used
in the validate_order_stock_available order validator.