Allow setting PRICE_INCLUDES_TAX per request/order
Fixes #55

This adds a new field to the Order model, you might need
to create a migration within your own app.

It also removes the ``unit_price`` property on the
PriceBase model, since there's no way for the price to
know it. You should only call unit_price on OrderItem.